Vaccine.py
The "Vaccine Booking System" is a simple program that allows users to:  

1. Book a vaccination.  
2. Search vaccination details.  
3. Delete records.  
4. View all records.  

All data is stored securely in a file, ensuring it is saved permanently and easily accessible. With digital storage, the system is more efficient, reliable, and convenient.  

This tool helps track the number of vaccinated individuals nationwide and can be crucial for monitoring in case of a future pandemic.